+++ ExchangeRatesBot.Domain/Interfaces/IUserService.cs
@@ -12,6 +12,6 @@ namespace ExchangeRatesBot.Domain.Interfaces
     {
         Task<bool> SetUser(long chatId, User user = default, CancellationToken cancel = default);
         Task<bool> Create(User user, CancellationToken cancel);
-        Task<bool> SubscribeUpdate(int chatId, bool subscribe, CancellationToken cancel);
+        Task<bool> SubscribeUpdate(long chatId, bool subscribe, CancellationToken cancel);
     }
 }
